+++ db/pvault.sql
@@ -82,6 +82,7 @@ CREATE TABLE `cart` (
  `id` int(11) NOT NULL AUTO_INCREMENT,
  `name` varchar(255) COLLATE utf8_bin NOT NULL,
  `quantity` varchar(255) COLLATE utf8_bin NOT NULL,
+ `purity` varchar(255) COLLATE utf8_bin NOT NULL,
  `supplier` varchar(255) COLLATE utf8_bin NOT NULL,
  `supplier_link` varchar(255) COLLATE utf8_bin NOT NULL, 
  PRIMARY KEY (`id`)